# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_internal_domain_model.Workspace:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      logoUrl:
        type: string
      name:
        type: string
      plan:
        type: string
      role:
        description: Current user's role in this workspace
        type: string
      slug:
        type: string
      updatedAt:
        type: string
    type: object
...
  backend_internal_presentation_response.Envelope:
    properties:
      data: {}
    type: object
  backend_internal_presentation_response.ErrorEnvelope:
    properties:
      title:
        type: string
    type: object
...
  backend_internal_presentation_workspace_presenter.CreateWorkspaceRequest:
    properties:
      logoUrl:
        type: string
      name:
        type: string
      plan:
        type: string
    required:
    - name
    type: object
...
info:
  contact: {}
  description: API documentation for Dokudocs backend service.
  title: Dokudocs API
  version: "1.0"
paths:
...
  /workspaces:
    get:
      description: Get list of workspaces accessible by current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/backend_internal_domain_model.Workspace'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List workspaces
      tags:
      - Workspace
    post:
      consumes:
      - application/json
      description: Create a new workspace for the authenticated user
      parameters:
      - description: Workspace creation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_internal_presentation_workspace_presenter.CreateWorkspaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_domain_model.Workspace'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create workspace
      tags:
      - Workspace
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"